refactor(order/repository): extract transaction logic into WithTx helper
- Introduced WithTx helper function to encapsulate transaction management
- Refactored AddOrder and DeleteOrder methods to use WithTx
