+++ order/internal/repository/order/postgres/create.go
@@ -2,78 +2,59 @@ package postgres
 
 import (
 	"context"
-	"fmt"
-	"log"
 
 	sq "github.com/Masterminds/squirrel"
+	"github.com/jackc/pgx/v5"
 
 	"github.com/andredubov/rocket-factory/order/internal/model"
 	"github.com/andredubov/rocket-factory/order/internal/repository/converter"
 )
 
 func (r *ordersRepository) AddOrder(ctx context.Context, order model.Order) error {
-	repoOrder := converter.OrderToRepoModel(order)
-	tx, err := r.pool.Begin(ctx)
-	if err != nil {
-		return fmt.Errorf("failed to begin transaction: %w", err)
-	}
+	return WithTx(ctx, r.pool, func(tx pgx.Tx) error {
+		repoOrder := converter.OrderToRepoModel(order)
 
-	committed := false
-	defer func() {
-		if !committed {
-			if err := tx.Rollback(ctx); err != nil {
-				log.Printf("failed to rollback transaction: %v", err)
-			}
-		}
-	}()
-
-	orderBuilderInsert := sq.Insert(OrdersTable).
-		PlaceholderFormat(sq.Dollar).
-		Columns(
-			UserUUIDTableColumn,
-			TotalPriceTableColumn,
-			TransactionUUIDTableColumn,
-			PaymentMethodTableColumn,
-			StatusTableColumn,
-		).
-		Values(
-			repoOrder.UserUUID,
-			repoOrder.TotalPrice,
-			repoOrder.PaymentInfo.TransactionUUID,
-			repoOrder.PaymentInfo.PaymentMethod,
-			repoOrder.Status,
-		)
-
-	orderQuery, orderQueryArgs, err := orderBuilderInsert.ToSql()
-	if err != nil {
-		return err
-	}
-
-	if _, err := tx.Exec(ctx, orderQuery, orderQueryArgs...); err != nil {
-		return err
-	}
-
-	for _, partUUID := range order.PartUUIDs {
-		orderPartsBuilderInsert := sq.Insert(OrderPartsTable).
+		orderBuilderInsert := sq.Insert(OrdersTable).
 			PlaceholderFormat(sq.Dollar).
-			Columns(OrderUUIDTableColumn, PartUUIDTableColumn).
-			Values(repoOrder.OrderUUID, partUUID)
+			Columns(
+				UserUUIDTableColumn,
+				TotalPriceTableColumn,
+				TransactionUUIDTableColumn,
+				PaymentMethodTableColumn,
+				StatusTableColumn,
+			).
+			Values(
+				repoOrder.UserUUID,
+				repoOrder.TotalPrice,
+				repoOrder.PaymentInfo.TransactionUUID,
+				repoOrder.PaymentInfo.PaymentMethod,
+				repoOrder.Status,
+			)
 
-		orderPartsQuery, orderPartsQueryArgs, err := orderPartsBuilderInsert.ToSql()
+		orderQuery, orderQueryArgs, err := orderBuilderInsert.ToSql()
 		if err != nil {
 			return err
 		}
 
-		if _, err := tx.Exec(ctx, orderPartsQuery, orderPartsQueryArgs...); err != nil {
+		if _, err := tx.Exec(ctx, orderQuery, orderQueryArgs...); err != nil {
 			return err
 		}
-	}
 
-	if err := tx.Commit(ctx); err != nil {
-		return fmt.Errorf("failed to commit transaction: %w", err)
-	}
+		for _, partUUID := range order.PartUUIDs {
+			orderPartsBuilderInsert := sq.Insert(OrderPartsTable).
+				PlaceholderFormat(sq.Dollar).
+				Columns(OrderUUIDTableColumn, PartUUIDTableColumn).
+				Values(repoOrder.OrderUUID, partUUID)
 
-	committed = true
+			orderPartsQuery, orderPartsQueryArgs, err := orderPartsBuilderInsert.ToSql()
+			if err != nil {
+				return err
+			}
 
-	return nil
+			if _, err := tx.Exec(ctx, orderPartsQuery, orderPartsQueryArgs...); err != nil {
+				return err
+			}
+		}
+		return nil
+	})
 }
