+++ order/internal/repository/order/postgres/repository.go
@@ -2,6 +2,8 @@ package postgres
 
 import (
 	"context"
+	"fmt"
+	"log"
 
 	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgconn"
@@ -43,4 +45,32 @@ func NewOrderRepository(pool PgxPool) service.OrdersRepository {
 	return &ordersRepository{
 		pool: pool,
 	}
+}
+
+// WithTx ...
+func WithTx(ctx context.Context, pool PgxPool, action func(tx pgx.Tx) error) error {
+	committed := false
+	tx, err := pool.Begin(ctx)
+	if err != nil {
+		return fmt.Errorf("failed to begin transaction: %w", err)
+	}
+
+	defer func() {
+		if !committed {
+			if err := tx.Rollback(ctx); err != nil {
+				log.Printf("failed to rollback transaction: %v", err)
+			}
+		}
+	}()
+
+	if err = action(tx); err != nil {
+		return err
+	}
+
+	if err := tx.Commit(ctx); err != nil {
+		return fmt.Errorf("failed to commit transaction: %w", err)
+	}
+
+	committed = true
+	return nil
 }
